requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.1.5

django-allauth==65.7.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests-oauthlib>=1.3.1

django-bootstrap5==24.2
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-redis==5.4.0
django-secure==1.0.2
django-storages==1.14.3
flake8==6.1.0
gunicorn==21.2.0
idna==3.10
importlib-metadata==4.11.4
iniconfig==2.1.0
Markdown==3.3.4
martor==1.6.26
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
python-markdown-math==0.9
redis==6.1.0
requests==2.32.3
sentry-sdk==1.42.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.6.0
zipp==3.21.0

# Тестирование
pytest==7.4.3
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.12.0
pytest-mock==3.12.0

# ИИ-помощник
google-generativeai>=0.8.0